% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DMC.R
\name{get_DMC}
\alias{get_DMC}
\title{Predict DMC by the LIS-based multiple testing procedure, where LIS values are predicted from the trained CRF model.}
\usage{
get_DMC(dt, alpha = 0.01)
}
\arguments{
\item{dt}{Data matrix resulted from load_data(), containing CRF features.}

\item{alpha}{Nominal FDR level.}
}
\value{
Data matrix containing raw MeDIP-seq, MRE-seq read counts and predicted DMC label.
}
\description{
Predict DMC by the LIS-based multiple testing procedure, where LIS values are predicted from the trained CRF model.
}
