% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DMR.R
\name{get_DMR}
\alias{get_DMR}
\title{Merge nearby DMCs to report DMRs, using heuristic from DSS method.}
\usage{
get_DMR(dt, minlen = 100, pct.sig = 0.7, minCG = 10)
}
\arguments{
\item{dt}{Data matrix resulted from get_DMC(), containing raw MeDIP-seq, MRE-seq read counts and predicted DMC label.}

\item{minlen}{Minimum length (in basepairs) required for DMR. Default is 100 bps.}

\item{minCG}{Minimum number of CpG sites required for DMR. Default is 10.}

\item{minlen}{In all DMRs, the percentage of CG sites with significant p-values (less than p.threshold) must be greater than this threshold. Default is 0.7. This is mainly used for correcting the effects of merging of nearby DMRs.}
}
\value{
A matrix of reported DMRs, specifying positions, total length and number of CpG sites contained.
}
\description{
Merge nearby DMCs to report DMRs, using heuristic from DSS method.
}
