% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load the data (MeDIP-seq, MRE-seq) and generate corresponding features.}
\usage{
load_data(ref_gene_path, medip1_path, medip2_path, mre1_path, mre2_path)
}
\arguments{
\item{ref_gene_path}{Path of the reference gene file, which is a bed file with 3 columns (chr, start, end) specifying the CpG sites under analysis.}

\item{medip1_path}{Path of the MeDIP-seq data for the first sample.}

\item{medip2_path}{Path of the MeDIP-seq data for the second sample.}

\item{mre1_path}{Path of the MRE-seq data for the first sample.}

\item{mre2_path}{Path of the MRE-seq data for the second sample.}
}
\value{
Data matrix containing squencing data features and genomic features, ready to be fitted into CRF model.
}
\description{
Load the data (MeDIP-seq, MRE-seq) and generate corresponding features.
}
