% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Train the CRF model with subset of DMC labels.}
\usage{
train(dt, train_path, c = 0.1, e = 0.005)
}
\arguments{
\item{dt}{Data matrix resulted from load_data(), containing CRF features.}

\item{train_path}{Path of the training DMC labels, which is a bed file containg a subset of dt where DMC labels are available.}

\item{c}{Hyper-parameter for the CRF. With larger c value, CRF tends to overfit the training data.}

\item{e}{Hyper-parameter for the CRF specifiying the termination criterion.}
}
\description{
Train the CRF model with subset of DMC labels.
}
